requirements.txt
# ────────────────
# 🌿 Core Framework
# ────────────────
kivy[base]>=2.3.0

# ────────────────
# ⚙️ Essential Buildozer & Python-for-Android Support
# ────────────────
appdirs
packaging
colorama>=0.3.3
jinja2
toml
virtualenv

# ────────────────
# 🌐 Networking Libraries
# ────────────────
requests
urllib3
pyOpenSSL

# ────────────────
# 🖥️ Desktop WebView
# ────────────────
pywebview

# ────────────────
# 🧪 Development & Testing
# ────────────────
pytest
pexpect

# ────────────────
# 🖥️ System Compatibility
# ────────────────
distlib
filelock
platformdirs
